requirements.txt
langchain-astradb
langchain
langchain_community
datasets
pypdf
python-dotenv
flask
sentence-transformers
langchain-ollama
celery
redis
gunicorn

-e .